Refactor multiplayer UI and fix lobby creation
