+++ src/app/ui/events.js
@@ -8,6 +8,7 @@ import {
   confirmPendingAction,
   cancelPendingAction,
   resolveCombat,
+  endTurn,
 } from '../game/core/index.js';
 import {
   handleHandCardClick,
@@ -15,260 +16,14 @@ import {
   activateCreatureAbility,
   handleLifeOrbClick,
 } from '../game/interactions.js';
-import { COLORS } from '../../game/cards/index.js';
 import {
-  addMultiplayerLogEvent,
-  isMultiplayerMatchActive,
-  seedMultiplayerMatch,
-  subscribeToMatch,
-  clearMatch,
-  MULTIPLAYER_EVENT_TYPES,
-} from '../multiplayer/runtime.js';
-import { generateId } from '../utils/id.js';
-
-const LOBBY_QUERY_LIMIT = 20;
-const STALE_LOBBY_TIMEOUT_MS = 60_000;
-const LOBBY_CLEANUP_THROTTLE_MS = 5_000;
-
-let lastLobbyCleanupCheck = 0;
-
-function ensureMultiplayerScreenSubscriptions() {
-  if (state.screen === 'multiplayer-lobbies' && !state.multiplayer.lobbySubscription) {
-    refreshLobbySubscription();
-  }
-  if (state.screen === 'multiplayer-lobby-detail' && state.multiplayer.activeLobby?.id) {
-    ensureActiveLobbySubscription(state.multiplayer.activeLobby.id);
-  }
-}
-
-function ensureActiveLobbySubscription(lobbyId) {
-  if (!lobbyId) return;
-  if (state.multiplayer.activeLobbySubscription) {
-    state.multiplayer.activeLobbySubscription();
-    state.multiplayer.activeLobbySubscription = null;
-  }
-
-  const query = {
-    lobbies: {
-      $: {
-        where: { id: lobbyId },
-        limit: 1,
-      },
-    },
-  };
-
-  const unsubscribe = db.subscribeQuery(query, (snapshot) => {
-    if (snapshot.error) {
-      state.multiplayer.activeLobby = null;
-      state.multiplayer.activeLobbySubscription = null;
-      clearMatch();
-      requestRender();
-      return;
-    }
-    const previousLobby = state.multiplayer.activeLobby;
-    const lobby = snapshot.data?.lobbies?.[0] ?? null;
-    state.multiplayer.activeLobby = lobby;
-    if (!lobby) {
-      state.multiplayer.activeLobbySubscription = null;
-      clearMatch();
-      if (state.screen === 'multiplayer-lobby-detail') {
-        state.screen = 'multiplayer-lobbies';
-      }
-      requestRender();
-      return;
-    }
-
-    const matchId = lobby.matchId ?? null;
-    if (matchId && matchId !== state.multiplayer.currentMatchId) {
-      subscribeToMatch(matchId);
-    } else if (!matchId && state.multiplayer.currentMatchId) {
-      clearMatch();
-    }
-
-    const userId = state.auth.user?.id;
-    const userInLobby = Boolean(userId && (lobby.hostUserId === userId || lobby.guestUserId === userId));
-    if (matchId && userInLobby) {
-      state.screen = 'game';
-    } else if (!matchId && previousLobby?.matchId && userInLobby && state.screen === 'game') {
-      state.screen = 'multiplayer-lobby-detail';
-    }
-    requestRender();
-  });
-
-  state.multiplayer.activeLobbySubscription = unsubscribe;
-}
-
-function cleanupActiveLobbySubscription() {
-  if (typeof state.multiplayer.activeLobbySubscription === 'function') {
-    state.multiplayer.activeLobbySubscription();
-  }
-  state.multiplayer.activeLobbySubscription = null;
-}
-
-function cleanupLobbyListSubscription() {
-  if (typeof state.multiplayer.lobbySubscription === 'function') {
-    state.multiplayer.lobbySubscription();
-  }
-  state.multiplayer.lobbySubscription = null;
-}
-
-function returnToModeSelectFromLobbies() {
-  cleanupActiveLobbySubscription();
-  clearMatch();
-  cleanupLobbyListSubscription();
-  state.multiplayer.activeLobby = null;
-  state.multiplayer.lobbyList.lobbies = [];
-  state.multiplayer.lobbyList.loading = false;
-  state.multiplayer.lobbyList.error = null;
-  state.multiplayer.lobbyList.searchTerm = '';
-  state.screen = 'mode-select';
-  requestRender();
-}
-
-function bindMultiplayerLobbyEvents(root) {
-  if (root.__multiplayerLobbyHandlers) return;
-
-  const handleClick = async (event) => {
-    const target = event.target.closest('[data-action]');
-    if (!target || !root.contains(target)) return;
-
-    const action = target.getAttribute('data-action');
-
-    if (action === 'back-mode-select' && state.screen === 'multiplayer-lobbies') {
-      event.preventDefault();
-      returnToModeSelectFromLobbies();
-      return;
-    }
-
-    if (action === 'clear-search' && state.screen === 'multiplayer-lobbies') {
-      event.preventDefault();
-      state.multiplayer.lobbyList.searchTerm = '';
-      refreshLobbySubscription();
-      requestRender();
-      return;
-    }
-
-    if (action === 'create-lobby' && state.screen === 'multiplayer-lobbies') {
-      event.preventDefault();
-      await createLobby();
-      return;
-    }
-
-    if (action === 'view-lobby' && state.screen === 'multiplayer-lobbies') {
-      event.preventDefault();
-      const lobbyId = target.getAttribute('data-lobby');
-      if (!lobbyId) return;
-      const lobby = state.multiplayer.lobbyList.lobbies.find((l) => l.id === lobbyId);
-      state.multiplayer.activeLobby = lobby || null;
-      state.screen = 'multiplayer-lobby-detail';
-      ensureActiveLobbySubscription(lobbyId);
-      if (lobby?.matchId) {
-        subscribeToMatch(lobby.matchId);
-      }
-      requestRender();
-      return;
-    }
-
-    if (action === 'back-lobbies' && state.screen === 'multiplayer-lobby-detail') {
-      event.preventDefault();
-      const lobby = state.multiplayer.activeLobby;
-      const userId = state.auth.user?.id;
-      if (lobby && userId) {
-        if (lobby.guestUserId === userId) {
-          await leaveSeat('guest');
-        }
-      }
-      cleanupActiveLobbySubscription();
-      clearMatch();
-      state.multiplayer.activeLobby = null;
-      state.screen = 'multiplayer-lobbies';
-      requestRender();
-      return;
-    }
-
-    if (action === 'claim-seat' && state.screen === 'multiplayer-lobby-detail') {
-      event.preventDefault();
-      const seat = target.getAttribute('data-seat');
-      if (!seat) return;
-      await claimSeat(seat);
-      return;
-    }
-
-    if (action === 'leave-seat' && state.screen === 'multiplayer-lobby-detail') {
-      event.preventDefault();
-      const seat = target.getAttribute('data-seat');
-      if (!seat) return;
-      await leaveSeat(seat);
-      return;
-    }
-
-    if (action === 'choose-deck' && state.screen === 'multiplayer-lobby-detail') {
-      event.preventDefault();
-      const seat = target.getAttribute('data-seat');
-      const color = target.getAttribute('data-color');
-      if (!seat || !color) return;
-      await chooseDeck(seat, color);
-      return;
-    }
-
-    if (action === 'toggle-ready' && state.screen === 'multiplayer-lobby-detail') {
-      event.preventDefault();
-      const seat = target.getAttribute('data-seat');
-      if (!seat) return;
-      await toggleReady(seat);
-      return;
-    }
-
-    if (action === 'start-match' && state.screen === 'multiplayer-lobby-detail') {
-      event.preventDefault();
-      await startMatch();
-      return;
-    }
-  };
-
-  const handleInput = (event) => {
-    const target = event.target;
-    if (!(target instanceof HTMLInputElement)) return;
-    if (target.getAttribute('data-action') === 'search-lobbies' && state.screen === 'multiplayer-lobbies') {
-      state.multiplayer.lobbyList.searchTerm = target.value ?? '';
-      requestRender();
-    }
-  };
-
-  const handleChange = (event) => {
-    const target = event.target;
-    if (!(target instanceof HTMLInputElement)) return;
-    if (target.getAttribute('data-action') === 'search-lobbies' && state.screen === 'multiplayer-lobbies') {
-      refreshLobbySubscription();
-    }
-  };
-
-  const handleKeydown = (event) => {
-    const target = event.target;
-    if (!(target instanceof HTMLInputElement)) return;
-    if (target.getAttribute('data-action') === 'search-lobbies' && state.screen === 'multiplayer-lobbies') {
-      if (event.key === 'Enter') {
-        event.preventDefault();
-        refreshLobbySubscription();
-      }
-    }
-  };
-
-  root.addEventListener('click', handleClick);
-  root.addEventListener('input', handleInput);
-  root.addEventListener('change', handleChange);
-  root.addEventListener('keydown', handleKeydown);
-
-  root.__multiplayerLobbyHandlers = {
-    handleClick,
-    handleInput,
-    handleChange,
-    handleKeydown,
-  };
-}
+  attachMultiplayerEventHandlers,
+  ensureMultiplayerScreenSubscriptions,
+  convertMatchToGame,
+} from './multiplayer/events.js';
 
 export function attachEventHandlers(root) {
-  bindMultiplayerLobbyEvents(root);
+  attachMultiplayerEventHandlers(root);
   ensureMultiplayerScreenSubscriptions();
   root.querySelectorAll('[data-action="start"]').forEach((btn) => {
     btn.addEventListener('click', () => {
@@ -558,184 +313,6 @@ function bindGameEvents(root) {
   );
 }
 
-async function claimSeat(seat) {
-  const user = state.auth.user;
-  const lobby = state.multiplayer.activeLobby;
-  if (!user || !lobby) return;
-
-  const seatUserKey = seat === 'host' ? 'hostUserId' : 'guestUserId';
-  const seatNameKey = seat === 'host' ? 'hostDisplayName' : 'guestDisplayName';
-  const seatReadyKey = seat === 'host' ? 'hostReady' : 'guestReady';
-  const seatColorKey = seat === 'host' ? 'hostColor' : 'guestColor';
-
-  if (lobby[seatUserKey] && lobby[seatUserKey] !== user.id) {
-    return;
-  }
-
-  const updates = {
-    updatedAt: Date.now(),
-    [seatUserKey]: user.id,
-    [seatNameKey]: deriveDisplayName(user).trim() || 'Player',
-    [seatReadyKey]: false,
-    [seatColorKey]: typeof lobby[seatColorKey] === 'string' ? lobby[seatColorKey] : '',
-  };
-
-  await db.transact(db.tx.lobbies[lobby.id].update(updates));
-}
-
-async function leaveSeat(seat) {
-  const lobby = state.multiplayer.activeLobby;
-  const userId = state.auth.user?.id;
-  if (!lobby || !userId) return;
-
-  const seatUserKey = seat === 'host' ? 'hostUserId' : 'guestUserId';
-  if (lobby[seatUserKey] !== userId) return;
-
-  const updates = {
-    updatedAt: Date.now(),
-    [seatUserKey]: '',
-    [seat === 'host' ? 'hostDisplayName' : 'guestDisplayName']: '',
-    [seat === 'host' ? 'hostColor' : 'guestColor']: '',
-    [seat === 'host' ? 'hostReady' : 'guestReady']: false,
-  };
-
-  await db.transact(db.tx.lobbies[lobby.id].update(updates));
-}
-
-async function chooseDeck(seat, color) {
-  const lobby = state.multiplayer.activeLobby;
-  const userId = state.auth.user?.id;
-  if (!lobby || !userId || !COLORS[color]) return;
-
-  const seatUserKey = seat === 'host' ? 'hostUserId' : 'guestUserId';
-  const seatColorKey = seat === 'host' ? 'hostColor' : 'guestColor';
-  const seatReadyKey = seat === 'host' ? 'hostReady' : 'guestReady';
-
-  if (lobby[seatUserKey] !== userId) return;
-
-  const opponentColorKey = seat === 'host' ? lobby.guestColor : lobby.hostColor;
-  if (opponentColorKey === color) {
-    return;
-  }
-
-  const updates = {
-    updatedAt: Date.now(),
-    [seatColorKey]: color,
-    [seatReadyKey]: false,
-  };
-
-  await db.transact(db.tx.lobbies[lobby.id].update(updates));
-}
-
-async function toggleReady(seat) {
-  const lobby = state.multiplayer.activeLobby;
-  const userId = state.auth.user?.id;
-  if (!lobby || !userId) return;
-
-  const seatUserKey = seat === 'host' ? 'hostUserId' : 'guestUserId';
-  const seatReadyKey = seat === 'host' ? 'hostReady' : 'guestReady';
-  const seatColorKey = seat === 'host' ? 'hostColor' : 'guestColor';
-
-  if (lobby[seatUserKey] !== userId) return;
-  if (!lobby[seatColorKey]) return; // must choose a deck first
-
-  const updates = {
-    updatedAt: Date.now(),
-    [seatReadyKey]: !lobby[seatReadyKey],
-  };
-
-  await db.transact(db.tx.lobbies[lobby.id].update(updates));
-}
-
-async function startMatch() {
-  const lobby = state.multiplayer.activeLobby;
-  const userId = state.auth.user?.id;
-  if (!lobby || lobby.hostUserId !== userId) return;
-
-  const ready =
-    lobby.hostUserId &&
-    lobby.guestUserId &&
-    lobby.hostColor &&
-    lobby.guestColor &&
-    lobby.hostReady &&
-    lobby.guestReady;
-
-  if (!ready) return;
-
-  const matchId = generateId('match');
-  const eventId = generateId('matchEvent');
-  const now = Date.now();
-  const diceRolls = {
-    host: 1 + Math.floor(Math.random() * 6),
-    guest: 1 + Math.floor(Math.random() * 6),
-  };
-  let winner = null;
-  if (diceRolls.host !== diceRolls.guest) {
-    winner = diceRolls.host > diceRolls.guest ? 0 : 1;
-  }
-
-  const activePlayer = winner ?? 0;
-  const match = {
-    id: matchId,
-    lobbyId: lobby.id,
-    status: 'starting',
-    activePlayer,
-    turn: 1,
-    phase: 'main1',
-    dice: {
-      host: diceRolls.host,
-      guest: diceRolls.guest,
-      winner,
-    },
-    state: null,
-    pendingAction: null,
-    winner: null,
-    nextSequence: 2,
-    createdAt: now,
-    updatedAt: now,
-  };
-
-  const matchStartedEvent = {
-    id: eventId,
-    matchId,
-    sequence: 1,
-    type: MULTIPLAYER_EVENT_TYPES.MATCH_STARTED,
-    payload: {
-      turn: 1,
-      activePlayer,
-      phase: 'main1',
-      dice: { ...match.dice },
-    },
-    createdAt: now,
-  };
-
-  try {
-    state.multiplayer.lobbyList.error = null;
-    await db.transact([
-      db.tx.matches[matchId].update(match),
-      db.tx.matchEvents[eventId].update(matchStartedEvent),
-      db.tx.lobbies[lobby.id].update({
-        matchId,
-        status: 'starting',
-        updatedAt: now,
-      }),
-    ]);
-
-    state.multiplayer.activeLobby = {
-      ...lobby,
-      matchId,
-      status: 'starting',
-      updatedAt: now,
-    };
-    subscribeToMatch(matchId);
-    requestRender();
-  } catch (error) {
-    console.error('Failed to start match', error);
-    state.multiplayer.lobbyList.error = 'Could not start the match. Please try again.';
-    requestRender();
-  }
-}
-
 function positionAttackLines(root) {
   const attackLines = root.querySelectorAll('.attack-lines-svg .attack-line');
   if (!attackLines.length) return;
@@ -879,160 +456,3 @@ function canCurrentUserAct() {
   return isActiveTurn || isPendingTarget || isBlockingTurn;
 }
 
-function maybeCleanupStaleLobbies(lobbies) {
-  const now = Date.now();
-  if (now - lastLobbyCleanupCheck < LOBBY_CLEANUP_THROTTLE_MS) {
-    return;
-  }
-
-  const userId = state.auth.user?.id ?? null;
-  let deletedAny = false;
-
-  for (const lobby of lobbies) {
-    if (!lobby) continue;
-    if (lobby.matchId) continue;
-
-    const status = lobby.status || 'open';
-    if (status !== 'open' && status !== 'ready') continue;
-
-    const hasGuest = Boolean(lobby.guestUserId);
-    if (hasGuest) continue;
-
-    const lastUpdated = lobby.updatedAt ?? lobby.createdAt ?? 0;
-    if (!lastUpdated) continue;
-
-    if (now - lastUpdated < STALE_LOBBY_TIMEOUT_MS) continue;
-
-    const canDelete = Boolean(!lobby.hostUserId || (userId && lobby.hostUserId === userId));
-    if (!canDelete) continue;
-
-    db
-      .transact(db.tx.lobbies[lobby.id].delete())
-      .catch((error) => console.error('Failed to delete stale lobby', lobby.id, error));
-    deletedAny = true;
-  }
-
-  if (deletedAny) {
-    lastLobbyCleanupCheck = now;
-  }
-}
-
-function refreshLobbySubscription() {
-  cleanupLobbyListSubscription();
-
-  state.multiplayer.lobbyList.loading = true;
-  state.multiplayer.lobbyList.error = null;
-  requestRender();
-
-  const query = {
-    lobbies: {
-      $: {
-        where: {
-          status: { $in: ['open', 'ready', 'starting', 'playing'] },
-        },
-        orderBy: [
-          { field: 'status', direction: 'asc' },
-          { field: 'updatedAt', direction: 'desc' },
-        ],
-        limit: LOBBY_QUERY_LIMIT,
-      },
-    },
-  };
-
-  const unsubscribe = db.subscribeQuery(query, (snapshot) => {
-    if (snapshot.error) {
-      state.multiplayer.lobbyList.loading = false;
-      state.multiplayer.lobbyList.error = snapshot.error.message || 'Failed to load lobbies.';
-      state.multiplayer.lobbyList.lobbies = [];
-      requestRender();
-      return;
-    }
-
-    const searchTerm = state.multiplayer.lobbyList.searchTerm.trim().toLowerCase();
-    const snapshotLobbies = snapshot.data?.lobbies ?? [];
-    maybeCleanupStaleLobbies(snapshotLobbies);
-
-    let lobbies = snapshotLobbies;
-    if (searchTerm) {
-      lobbies = lobbies.filter((lobby) => {
-        const host = (lobby.hostDisplayName || '').toLowerCase();
-        const guest = (lobby.guestDisplayName || '').toLowerCase();
-        return host.includes(searchTerm) || guest.includes(searchTerm);
-      });
-    }
-
-    state.multiplayer.lobbyList.lobbies = lobbies;
-    state.multiplayer.lobbyList.loading = false;
-    requestRender();
-  });
-
-  state.multiplayer.lobbySubscription = unsubscribe;
-}
-
-async function createLobby() {
-  const user = state.auth.user;
-  if (!user) {
-    state.multiplayer.lobbyList.error = 'You must be signed in to create a lobby.';
-    requestRender();
-    return;
-  }
-
-  try {
-    const lobbyId = generateId('lobby');
-    const now = Date.now();
-    const displayName = deriveDisplayName(user);
-    const normalizedName = displayName.trim() || 'Player';
-    state.multiplayer.lobbyList.error = null;
-    const lobby = {
-      id: lobbyId,
-      status: 'open',
-      hostUserId: user.id,
-      hostDisplayName: normalizedName,
-      hostColor: '',
-      hostReady: false,
-      guestUserId: '',
-      guestDisplayName: '',
-      guestColor: '',
-      guestReady: false,
-      searchKey: normalizedName.toLowerCase(),
-      matchId: '',
-      createdAt: now,
-      updatedAt: now,
-    };
-
-    await db.transact(db.tx.lobbies[lobbyId].update(lobby));
-    cleanupActiveLobbySubscription();
-    state.multiplayer.activeLobby = lobby;
-    state.screen = 'multiplayer-lobby-detail';
-    ensureActiveLobbySubscription(lobbyId);
-    requestRender();
-  } catch (error) {
-    console.error('Failed to create lobby', error);
-    state.multiplayer.lobbyList.error = 'Could not create lobby. Please try again.';
-    requestRender();
-  }
-}
-
-function convertMatchToGame() {
-  if (!state.multiplayer.match) return;
-  state.game = null;
-  state.screen = 'game';
-  requestRender();
-}
-
-function deriveDisplayName(user) {
-  if (!user) return 'Player';
-  if (typeof user.displayName === 'string' && user.displayName.trim()) {
-    return user.displayName.trim();
-  }
-  if (typeof user.username === 'string' && user.username.trim()) {
-    return user.username.trim();
-  }
-  if (typeof user.name === 'string' && user.name.trim()) {
-    return user.name.trim();
-  }
-  if (typeof user.email === 'string' && user.email.includes('@')) {
-    return user.email.split('@')[0];
-  }
-  return 'Player';
-}
